requirements.txt
python-dotenv
langchain==0.3.1
langchain-core==0.3.7
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-mistralai==0.2.0
langchain-qdrant==0.1.4
langgraph==0.2.32
feedparser
tavily-python